+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+2019-12-05
+----------
+
+ * Add `collectionClass` to EntityLocator to make it easier to override
+
 2019-12-03
 ----------